More

+++ pyrecest/evaluation/evaluate_for_simulation_config.py
@@ -61,7 +61,7 @@ def evaluate_for_simulation_config(
 
 def get_all_seeds(n_runs: int, seed_input=None, consecutive_seed: bool = True):
     if seed_input is None:
-        seed_input = random.randint(1, 0xFFFFFFFF, dtype=uint32)  # nosec
+        seed_input = random.randint(1, 0xFFFFFFFF)  # nosec
 
     if seed_input.shape[0] == n_runs:
         all_seeds = seed_input
